backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
pip==24.0
psycopg2==2.9.9
setuptools==68.0.0
djangorestframework-simplejwt==5.3.1
django_cors_headers==4.3.1
drf-spectacular==0.27.1
django-filter==24.2